Cleaned up the hit testing a little bit
I still have a ways to go to clean up the code, but now we've got hit testing modulated into the PhysicalObject class at the method level. Now objects have to be drawn as triangle strips in order to make tesselation of larger shapes easier. Needs a shit load of documentation which is the next thing I'll do.